requirements.txt
Django==5.1.2
mysqlclient==2.2.5
django-environ==0.11.2
djangorestframework==3.15.2
pillow==11.0.0
drf-yasg==1.21.8
django-filter==24.3
django-silk==5.3.1
djangorestframework-simplejwt==5.3.1
celery==5.4.0
redis==5.2.0